produto-pipeline/requirements.txt
crewai>=0.80.0
crewai-tools>=0.15.0
llama-index>=0.10.0
llama-index-embeddings-openai>=0.1.0
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
slack-sdk>=3.27.0
PyGithub>=2.1.0
ruff>=0.6.0
mypy>=1.11.0
pytest>=8.0.0
pytest-cov>=5.0.0
